add generation

+++ src/main/java/org/project/hackathonteamgenerator/repository/ParticipantRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParticipantRepository extends CrudRepository<Participant, Integer> {
+public interface ParticipantRepository extends CrudRepository<Participant, Long> {
 }
